chore(slack): rename shared app environment module
