+++ apps/sim/lib/slack-search/shared-app-env.ts
